requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
python-dotenv
gradio
plotly
dash>=3.0.0
dash-bootstrap-components>=2.0.0
flask>=3.0.0
alpaca-py>=0.8.2
